feat: upgrade SvelteKit, merge configs
