+++ package.json
@@ -7,11 +7,11 @@
   "scripts": {
     "packages": "pnpm --filter \"./packages/**\"",
     "start": "node build/server.js",
-    "dev": "svelte-kit dev",
+    "dev": "vite dev",
     "dev:lib": "pnpm -r run build -w",
-    "build": "svelte-kit build && pnpm build:server",
+    "build": "vite build && pnpm build:server",
     "build:lib": "pnpm packages run build",
-    "build:server": "vite build",
+    "build:server": "vite build --mode server",
     "package": "svelte-kit package",
     "preview": "svelte-kit preview",
     "prepare": "prisma generate",
@@ -37,11 +37,11 @@
     "@hey-amplify/scripts": "workspace:*",
     "@hey-amplify/support": "workspace:*",
     "@hey-amplify/tsconfig": "workspace:*",
-    "@sveltejs/adapter-node": "1.0.0-next.76",
-    "@sveltejs/kit": "1.0.0-next.345",
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.41",
+    "@sveltejs/adapter-node": "1.0.0-next.79",
+    "@sveltejs/kit": "1.0.0-next.366",
     "@typescript-eslint/eslint-plugin": "^5.26.0",
     "@typescript-eslint/parser": "^5.26.0",
+    "@vitest/ui": "^0.17.1",
     "carbon-components-svelte": "^0.64.2",
     "carbon-icons-svelte": "^11.0.1",
     "carbon-preprocess-svelte": "^0.9.0",
@@ -52,15 +52,13 @@
     "prettier": "^2.6.2",
     "prettier-plugin-svelte": "^2.7.0",
     "prisma": "^3.14.0",
-    "svelte": "3.48.0",
-    "svelte-check": "^2.7.1",
-    "svelte-preprocess": "^4.10.6",
-    "svelte2tsx": "^0.5.10",
+    "svelte": "3.49.0",
+    "svelte-check": "^2.8.0",
+    "svelte-preprocess": "^4.10.7",
     "tslib": "^2.4.0",
     "typescript": "^4.7.2",
-    "vite": "^2.9.9",
-    "vite-node": "^0.12.9",
-    "vitest": "^0.12.9"
+    "vite": "^2.9.14",
+    "vitest": "^0.17.1"
   },
   "prettier": {
     "tabWidth": 2,
